mcp-playwright-min/requirements.txt
playwright==1.43.0
mcp==1.3.0
python-dotenv==1.0.1
